+++ pyweatherfr/__init__.py
@@ -8,16 +8,16 @@ from pyweatherfr.pyweatherfr import find
 from pyweatherfr.update import update
 import colorama
 
+
 def pyweatherfr():
     """
     pyweatherfr entry point
     """
-    args = compute_args()
+    
     colorama.init()
-    if args.day and (args.town is None):
-        print("--day requires --town")
-        exit(1)
+
+    args = compute_args()
     if args.update:
         update()
-    if args.town:
-        find(args.town)
+    else:
+        find()