+++ pyweatherfr/args.py
@@ -21,7 +21,7 @@ def compute_args():
         There is NO WARRANTY, to the extent permitted by law.
         Written by thib1984.""",
     )
-    my_group = my_parser.add_mutually_exclusive_group(required=True)
+    my_group = my_parser.add_mutually_exclusive_group()
     my_group.add_argument(
         "town",
         metavar="town",
@@ -50,7 +50,13 @@ def compute_args():
         "--condensate",
         action="store_true",
         help="condensate sysout",
-    )       
+    )
+    my_parser.add_argument(
+        "-v",
+        "--verbose",
+        action="store_true",
+        help="debug mode",
+    )              
     my_group.add_argument(
         "-u",
         "--update",
@@ -58,10 +64,6 @@ def compute_args():
         help="update pyweatherfr",
     )
 
-    # if no parameter
-    if len(sys.argv) == 1:
-        my_parser.print_help()
-        sys.exit(0)
 
     args = my_parser.parse_args()
     return args